+++ raw_stdin_other.go
@@ -1,0 +1,12 @@
+//go:build !windows
+
+package main
+
+import (
+	"io"
+	"os"
+)
+
+func rawStdinReader() io.Reader {
+	return os.Stdin
+}